Rename namespace GHType to Type
